Some fixes

+++ slurm/fix_usd.py
@@ -9,14 +9,15 @@ def fix_usd_path(path):
     try:
         # make a backup
         backup_path = path + ".bak"
-        shutil.copy2(path, backup_path)
+        if not os.path.exists(backup_path):
+            shutil.copy2(path, backup_path)
 
         stage = Usd.Stage.Open(path)
         def _update_path(asset_path):
             if "fsx-siro" not in asset_path or "material" not in asset_path:
                 return asset_path
             fixed_path = os.path.join("..", "material", os.path.basename(asset_path))
-            print(asset_path, "->", fixed_path)
+            # print(asset_path, "->", fixed_path)
             return fixed_path
 
         UsdUtils.ModifyAssetPaths(stage.GetRootLayer(), _update_path)
@@ -27,7 +28,7 @@ def fix_usd_path(path):
 def main():
     with ProcessPoolExecutor() as executor:
         # Map the fix_usd_path function to all USD files
-        usds = list(glob.glob("/home/cgokmen/projects/behavior-data2/hssd/objects/*/*/usd/*.usd"))
+        usds = list(glob.glob("/fsx-siro/cgokmen/behavior-data2/*/objects/*/*/usd/*.usd"))
         assert usds
 
         futures = {executor.submit(fix_usd_path, usd): usd for usd in usds}
